fixup! Implement caching for VR, TTS, Buttons, VehicheInfo and RC capabilities
