+++ src/components/application_manager/test/include/application_manager/hmi_capabilities_for_testing.h
@@ -45,7 +45,7 @@ class HMICapabilitiesForTesting
   HMICapabilitiesForTesting(::application_manager::ApplicationManager& app_mngr)
       : HMICapabilitiesImpl(app_mngr) {}
   bool LoadCapabilitiesFromFile() {
-    return load_capabilities_from_file();
+    return LoadCapabilitiesFromFile();
   }
 };
